Restructure to cobra cmd/ layout, move m3u to internal/.
- Split subcommands into cmd/{root,organize,m3u,playlists}.go; main.go is now a thin cmd.Execute() entry point.
- Move m3u/ to internal/m3u/ since it has no external consumers.
- Scope --source/--destination required flags to organize where they belong (previously leaked onto every subcommand via the root).
- Ignore docs/ for local design notes.
- Refresh README example usage (the old cmus2m3u command no longer exists) and bump LICENSE copyright year.
