+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright 2024 Nick Granado <ngranado@gmail.com>
+Copyright 2026 Nick Granado <ngranado@gmail.com>
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
